Replaced convoluted instructions with actions
The reader is instructed to figure out if an action doesn't support a resource constraint by comparing the full list of actions (on another page) to the list in the table. The actions _not_ in the list do not support a resource constraint.

I removed those instructions and simply added 2 rows for the actions that _do not_ support the resource action. This makes it much easier for the read to know exactly which actions support the resource actions and which do not.