Update get-set-up-for-amazon-ecr.md
On line #3  "these" should be replaced with "those". For example, the following sentence uses those "If you attend high school A or high school B, you have a better shot of getting into Harvard because those two schools offer advanced placement classes." 
At the end of the sentence I used the word "both" rather than using "those" again. 

Also on #3 the "must" should be changed to "we recommend". This is because the CLI will still work for basic ECR features but it may not have commands for the latest features. As an alternative, the word "should" could replace "must". I chose "We recommend" because "should" is used in the next sentence. 

Thank you for taking the time to review my proposed changes.